CN_Project/app/api/requirements.txt
#setuptools==49.6.0
#grpcio-tools==1.46.5
#Protobuf==3.19.5
#connexion[swagger-ui]
#attrs==22.1.0
#certifi==2022.9.24
#charset-normalizer==2.1.1
#click==8.1.3
#clickclick==20.10.2
#Flask==2.2.2
#idna==3.4
#inflection==0.5.1
#itsdangerous==2.1.2
#Jinja2==3.1.2
#jsonschema==4.16.0
#MarkupSafe==2.1.1
#packaging==21.3
#pyparsing==3.0.9
#pyrsistent==0.19.1
#PyYAML==6.0
#requests==2.28.1
#swagger-ui-bundle==0.0.9
#urllib3==1.26.12
#Werkzeug==2.2.2
grpcio-tools >= 1.30
connexion[swagger-ui]
prometheus-client
#connexion >= 2.0.0